refactored model location
