+++ src/tradinghours/catalog.py
@@ -1,13 +1,13 @@
 from typing import Dict, Generator, Optional, Tuple, Type, TypeVar
 
-from tradinghours.validate import validate_str_arg
+from .validate import validate_str_arg
 
 from .base import BaseObject
-from .currency import Currency, CurrencyHoliday
-from .market import Market, MarketHoliday, MicMapping
+from .models.market import Market, MarketHoliday, MicMapping
+from .models.season import SeasonDefinition
+from .models.currency import CurrencyHoliday, Currency
+from .models.schedule import Schedule
 from .remote import default_data_manager
-from .schedule import Schedule
-from .season import SeasonDefinition
 from .store.engine import Store, default_store
 from .store.file import FileCollection
 from .store.source import SourceFile
