+++ src/tradinghours/models/season.py
@@ -1,8 +1,8 @@
 import datetime
 
-from .base import BaseObject, DateField, IntegerField, StringField, class_decorator
-from .exceptions import MissingDefinitionError
-from .validate import validate_int_arg, validate_str_arg
+from ..base import BaseObject, DateField, IntegerField, StringField, class_decorator
+from ..exceptions import MissingDefinitionError
+from ..validate import validate_int_arg, validate_str_arg
 
 @class_decorator
 class SeasonDefinition(BaseObject):
